backend/pyproject.toml
[tool.poetry]
name = "disc-golf-scorecard-backend"
version = "0.1.0"
description = "FUFA Disc Golf League — scorecard parser backend"
authors = ["Derek"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
anthropic = "^0.40.0"
gspread = "^6.1.0"
google-auth = "^2.36.0"
python-multipart = "^0.0.12"
python-dotenv = "^1.0.0"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
